requirements.txt
mecab-python3
ipadic
pandas
regex